🚚 Move CategorySelection component to single file
Now that the screen will have more components, it is a good idea to
break components in multiple files to make it easier to read.
