+++ features/task/src/main/java/com/escodro/task/presentation/detail/TaskDetail.kt
@@ -1,17 +1,11 @@
 package com.escodro.task.presentation.detail
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.preferredSize
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.TextField
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Create
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.outlined.Close
 import androidx.compose.runtime.Composable
@@ -19,12 +13,10 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import com.escodro.task.R
 import com.escodro.task.model.Category
 import com.escodro.task.model.Task
@@ -147,42 +139,6 @@ private fun TaskDescriptionTextField(text: String?, onDescriptionChanged: (Strin
         textStyle = MaterialTheme.typography.body1,
         backgroundColor = MaterialTheme.colors.background,
     )
-}
-
-@Composable
-internal fun CategorySelection(
-    categories: List<Category>,
-    currentCategory: Long?,
-    onCategoryChanged: (Long?) -> Unit
-) {
-    val currentItem = categories.find { category -> category.id == currentCategory }
-    val selectedState = remember { mutableStateOf(currentItem) }
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .preferredSize(56.dp)
-            .padding(start = 12.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        LeadingIcon(
-            imageVector = Icons.Default.Create,
-            contentDescription = R.string.task_detail_cd_icon_category
-        )
-        LazyRow(modifier = Modifier.padding(start = 16.dp)) {
-            items(
-                items = categories,
-                itemContent = { category ->
-                    val isSelected = category == selectedState.value
-                    CategoryItemChip(
-                        category = category,
-                        isSelected = isSelected,
-                        selectedState,
-                        onCategoryChanged = onCategoryChanged
-                    )
-                }
-            )
-        }
-    }
 }
 
 @Suppress("UndocumentedPublicFunction")
